+++ ionoscloud/data_source_datacenter.go
@@ -118,7 +118,7 @@ func dataSourceDataCenterRead(ctx context.Context, d *schema.ResourceData, meta 
 		}
 		log.Printf("[INFO] Got dc [Name=%s, Location=%s]", *datacenter.Properties.Name, *datacenter.Properties.Location)
 	} else {
-		datacenters, apiResponse, err := client.DataCentersApi.DatacentersGet(ctx).Execute()
+		datacenters, apiResponse, err := client.DataCentersApi.DatacentersGet(ctx).Depth(1).Execute()
 		logApiRequestTime(apiResponse)
 
 		if err != nil {